+++ tests/testthat/test-GetCovariates.R
@@ -60,6 +60,7 @@ test_that("getDbCovariateData enforces specification of database details", {
 })
 
 test_that("getDbCovariateData CDM v4 not supported", {
+  skip_on_cran()
   skip_if_not(dbms == "sqlite" && exists("eunomiaConnection"))
   expect_error(getDbCovariateData(
     connectionDetails = eunomiaConnectionDetails,
@@ -83,6 +84,7 @@ test_that("getDbCovariateData cohortTableIsTemp tests when table name lacks # sy
 })
 
 test_that("getDbCovariateData cohortTableIsTemp tests when table name contains # symbol", {
+  skip_on_cran()
   skip_if_not(dbms == "sqlite" && exists("eunomiaConnection"))
   result <- getDbCovariateData(
     connection = eunomiaConnection,
@@ -95,6 +97,7 @@ test_that("getDbCovariateData cohortTableIsTemp tests when table name contains #
 })
 
 test_that("getDbCovariateData populationSize == 0 tests", {
+  skip_on_cran()
   skip_if_not(dbms == "sqlite" && exists("eunomiaConnection"))
   expect_warning(getDbCovariateData(
     connection = eunomiaConnection,
@@ -107,6 +110,7 @@ test_that("getDbCovariateData populationSize == 0 tests", {
 })
 
 test_that("Custom covariate builder", {
+  skip_on_cran()
   # TODO: This test is probably good to run on all DB platforms
   skip_if_not(dbms == "sqlite" && exists("eunomiaConnection"))
   covariateSettings <- createCovariateSettings(