+++ tests/testthat/test-GetCovariatesTemporalSequence.R
@@ -55,6 +55,8 @@ test_that("getDbCovariateData works with createTemporalSequenceCovariateSettings
     useDemographicsAgeGroup = T,
     useConditionEraGroupStart = T,
     useDrugEraStart = T,
+    useMeasurement = T,
+    useMeasurementValue = T,
     timePart = "month",
     timeInterval = 1,
     sequenceEndDay = -1,
@@ -78,6 +80,7 @@ test_that("getDbCovariateData works with createTemporalSequenceCovariateSettings
 
 # Check backwards compatibility
 test_that("Temporal Covariate Settings are backwards compatible", {
+  skip_on_cran()
   skip_if_not(dbms == "sqlite" && exists("eunomiaConnection"))
 
   # Temporal covariate settings created previously will not have