+++ NEWS.md
@@ -1,3 +1,21 @@
+FeatureExtraction 3.7.2
+=======================
+
+- Fix bug in MeasurementValue.sql (#279)
+- Fix running getDbCovariateData on snowflake (#236)
+
+FeatureExtraction 3.7.1
+=======================
+
+- Fix bug related to minCharacterizationMean (#274)
+
+FeatureExtraction 3.7.0
+=======================
+
+- Deprecate oracleTempSchema param, add tempEmulationSchema param (#126)
+- Filter binary covariates server side (#255)
+- Limit tests that are run on CRAN (#267)
+
 FeatureExtraction 3.6.0
 =======================